fixed bug of lda.

+++ extends/src/main/java/ksc/mahout/lda/map/LDADocMapper.java
@@ -2,6 +2,7 @@ package ksc.mahout.lda.map;
 
 import ksc.mahout.util.Constant;
 import ksc.mahout.writable.UidPrefWritable;
+import org.apache.hadoop.io.IntWritable;
 import org.apache.hadoop.io.LongWritable;
 import org.apache.hadoop.io.Text;
 import org.apache.hadoop.mapreduce.Mapper;
@@ -12,11 +13,11 @@ import java.io.IOException;
 /**
  * Created by ZhangShengtao on 14-8-7.
  */
-public class LDADocMapper extends Mapper<LongWritable, VectorWritable, Text, UidPrefWritable> {
+public class LDADocMapper extends Mapper<IntWritable, VectorWritable, Text, UidPrefWritable> {
     private static String FLAG = "doc";
 
     @Override
-    protected void map(LongWritable key, VectorWritable value, Context context) throws IOException, InterruptedException {
+    protected void map(IntWritable key, VectorWritable value, Context context) throws IOException, InterruptedException {
         context.write(new Text(key.toString()), new UidPrefWritable(key.get(), Constant.FLAG_DOC, "0", value));
 
     }
